test case 생성 및 domain 별 repo, service 등 패키징 분리

+++ src/main/java/cc/domain/employee/entity/Employee.java
@@ -23,7 +23,8 @@ public class Employee {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "employee_id", nullable = false)
-    private Integer id;
+    private
+    Long id;
 
     @Size(max = 100)
     @NotNull
